Move shopping-cart-endpoint types to use-shopping-cart-manager
This centralizes the types of useShoppingCartManager and makes it easier
in the future to move this hook outside of checkout for use elsewhere.
