+++ client/my-sites/checkout/composite-checkout/add-items.ts
@@ -14,7 +14,7 @@ import {
 	jetpackProductItem,
 } from 'lib/cart-values/cart-items';
 import { JETPACK_PRODUCTS_LIST, JETPACK_SEARCH_PRODUCTS } from 'lib/products-values/constants';
-import type { RequestCartProduct } from './types/backend/shopping-cart-endpoint';
+import type { RequestCartProduct } from './hooks/use-shopping-cart-manager/types';
 import config from 'config';
 
 const debug = debugFactory( 'calypso:composite-checkout:add-items' );
