+++ client/my-sites/checkout/composite-checkout/components/secondary-cart-promotions.tsx
@@ -12,7 +12,7 @@ import config from 'config';
 import CartFreeUserPlanUpsell from 'my-sites/checkout/cart/cart-free-user-plan-upsell';
 import UpcomingRenewalsReminder from 'my-sites/checkout/cart/upcoming-renewals-reminder';
 import { getSelectedSiteId } from 'state/ui/selectors';
-import { ResponseCart } from '../types/backend/shopping-cart-endpoint';
+import { ResponseCart } from '../hooks/use-shopping-cart-manager/types';
 
 type PartialCart = Pick< ResponseCart, 'products' >;
 interface Props {
