+++ client/my-sites/checkout/composite-checkout/hooks/use-shopping-cart-manager/use-cart-update-and-revalidate.ts
@@ -7,12 +7,11 @@ import debugFactory from 'debug';
 /**
  * Internal dependencies
  */
-import { ResponseCart, RequestCart } from '../../types/backend/shopping-cart-endpoint';
 import {
 	convertResponseCartToRequestCart,
 	convertRawResponseCartToResponseCart,
 } from './cart-functions';
-import { CacheStatus, ShoppingCartAction } from './types';
+import { ResponseCart, RequestCart, CacheStatus, ShoppingCartAction } from './types';
 
 const debug = debugFactory( 'calypso:composite-checkout:use-cart-update-and-revalidate' );
