+++ client/my-sites/checkout/composite-checkout/hooks/use-shopping-cart-manager/use-initialize-cart-from-server.ts
@@ -7,9 +7,8 @@ import debugFactory from 'debug';
 /**
  * Internal dependencies
  */
-import { ResponseCart, RequestCart } from '../../types/backend/shopping-cart-endpoint';
 import { convertRawResponseCartToResponseCart } from './cart-functions';
-import { CacheStatus, ShoppingCartAction } from './types';
+import { ResponseCart, RequestCart, CacheStatus, ShoppingCartAction } from './types';
 
 const debug = debugFactory( 'calypso:composite-checkout:use-initialize-cart-from-server' );
